Reduzida quantidade de rotas para uma reserva. Agora clientes e não clientes usam as mesmas rotas para reservas

+++ model/Reserva.js
@@ -13,7 +13,7 @@ const Reserva = conexao.define('reserva', {
     },
     dataSaida: Sequelize.DATEONLY,
     dataChegada: Sequelize.DATEONLY,
-    valorDiaria: Sequelize.INTEGER,
+    valorDiaria: Sequelize.DOUBLE,
     diarias: Sequelize.INTEGER
     },{
         timestamps: false,
